Add New Python Files

+++ PDF Parsing.py
@@ -5,8 +5,9 @@ from bs4 import BeautifulSoup
 import numpy as np
 import os
 from urllib.parse import urljoin
+#from pdf2image import convert_from_path
 
-#import fitz
+import fitz
 
 def extract_information(pdf_path):
     with open(pdf_path, 'rb') as f:
@@ -34,9 +35,17 @@ def extract_information(pdf_path):
     
     return information
 
+def file_exists(file):
+    ind = os.path.exists(file)
+    return ind
+
 #This function takes a list search queries and searches the bioRxiv for relevant articles and downloads them to the environment
 def search_bioRxiv(queries):
+    page_num = 1
     for query in queries:
+        directory = query
+        path = os.path.join(floc, directory)
+        os.mkdir(path)
         query = query.replace(" ", "%252B")
         URL = "https://www.biorxiv.org/search/" + query + "%20jcode%3Abiorxiv%20numresults%3A75%20sort%3Arelevance-rank%20format_result%3Astandard"
         req = urllib.request.Request(URL, headers={'User-Agent' : "Magic Browser"})
@@ -51,10 +60,11 @@ def search_bioRxiv(queries):
         print(resultStats)
 
         #Retrieving page numbers
-        page_list = soup.find('div', class_='highwire-list page-group-items item-list')
-        pages = page_list.find_all('a')
-        page_num = int(pages[-1].get_text())
-        print(page_num)
+        if(soup.find('div', class_='highwire-list page-group-items item-list')):
+            page_list = soup.find('div', class_='highwire-list page-group-items item-list')
+            pages = page_list.find_all('a')
+            page_num = int(pages[-1].get_text())
+            print(page_num)
 
         #Accessing articles
         for x in range(0, page_num):
@@ -71,33 +81,67 @@ def search_bioRxiv(queries):
             for link in results:
                 new_URL = "https://www.biorxiv.org" + link['href'] + ".full.pdf"
                 #Name the pdf files using the last portion of each link which are unique in this case
-                filename = os.path.join(floc ,link['href'].split('/')[-1])
-                with open(filename, 'wb') as f:
-                    f.write(requests.get(new_URL).content)
+                filename = os.path.join(path, link['href'].split('/')[-1] + ".full.pdf")
+                # with open(filename, 'wb') as f:
+                #     f.write(requests.get(new_URL).content)
+                if file_exists(filename) == True:
+                    print("File Already Exists")
+                elif file_exists(filename) == False:
+                    with open(filename, 'wb') as f:
+                        f.write(requests.get(new_URL).content)
+
+def get_web_metadata(URL):
+    req = urllib.request.Request(URL, headers={'User-Agent' : "Magic Browser"})
+    response = urllib.request.urlopen( req )
+    html = response.read()
+
+    # Parsing response
+    soup = BeautifulSoup(html, 'html.parser')
+    
+    #Retrieve metadata from bioRxiv webpage
+    meta_data = soup.find('head')
+    meta = meta_data.find_all('meta')
+    print(meta)
+    
+
 
 if __name__ == '__main__':
    
     #File location that PDFs will be downloaded to in the virtual environment
-    #floc = "/home/dhunt/py-venv/2021-SULI-Text-Mining1/SULI-Text-Mining-Project"
-    floc = "/hpcgpfs01/scratch/dhuntSULI2021/2021-SULI-Text-Mining/SULI-Text-Mining-Project/bioRxiv_Articles"
+    floc = "/home/dhunt/py-venv/2021-SULI-Text-Mining1/SULI-Text-Mining-Project/bioRxiv_Articles/"
+    #floc = "/home/dhunt/py-venv/2021-SULI-Text-Mining1/SULI-Text-Mining-Project/bioRxiv_Articles2/"
+    #floc = "/hpcgpfs01/scratch/dhuntSULI2021/2021-SULI-Text-Mining/SULI-Text-Mining-Project/bioRxiv_Articles/"
 
     #search_queries = ["P-loop NTPase", "P-loop GTPase", "Fe protein", "Walker A motif", "nitrogenase reductase", "Nif-like", "dinitrogenase", "MoFe protein", "heterotetramer", "Molybdenum nitrogenase"]
-    search_queries = ["protein"]
+    #search_queries = ["fe AND protein AND crystal"]
+    search_queries = ["MoFe protein nitrogenase", "Molybdenum-iron protein nitrogenase", "MoFeP nitrogenase", "MoFe protein crystallization", "Molybdenum-iron protein crystallization", "MoFeP crystallization", "MoFe protein crystal structure", "Molybdenum-iron protein crystal structure", "MoFeP crystal structure", "MoFe protein crystallized", "Molybdenum-iron protein crystallized", "MoFeP crystallized"]
 
     search_bioRxiv(search_queries)
 
+    #srun -p volta -A nlp -sbu -t 24:00:00 -N1 --gres=gpu:8 -J test \
+    
     #Test PDF File Paths 
     # path = '2021.02.08.430359v2.full.pdf'
-    path = '439992v1'
+    path = 'bioRxiv_Articles/2021.07.01.450417v1.full.pdf'
 
     #Print metadata - PyPDF2.pdf
-    extract_information(path)
+    #extract_information(path)
+
+    #get_web_metadata("https://www.biorxiv.org/content/10.1101/2021.05.13.443919v2")
+
+    # Store Pdf with convert_from_path function
+    # images = convert_from_path(path)
+
+    # for i in range(len(images)):
+    #     # Save pages as images in the pdf
+    #     images[i].save('page'+ str(i) +'.jpg', 'JPEG'
     
     #Print metadata - PyMuPDF
     # doc = fitz.open(path)
-    # page = doc.load_page(0)
-    # met = doc.metadata
-    # print(met)
+    # page = doc[0]
+    # #met = doc.metadata
+    # links = page.get_toc()
+    # print(links)
 
     ##prints text on a page 
     #text = page.get_text("html")
